requirements.txt
streamlit==1.51.0
python-dotenv==1.1.1
requests==2.32.5
crewai==1.7.2
crewai-tools==1.7.2
litellm==1.75.3
pydantic==2.11.10
tavily-python==0.7.17